requirements.txt
rdkit-pypi
numpy
pandas
seaborn
matplotlib
requests
